pyproject.toml
[tool.poetry]
name = "github_repo_traffic"
version = "2.0.4"
description = ""
authors = ["Will Humphlett <will@humphlett.net>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.20.0"
PyGithub = "^1.55"
pandas = "^1.4.3"
asciichartpy = "^1.5.25"

[tool.poetry.dev-dependencies]
black = "^22.6.0"


[tool.poetry.scripts]
repo-data = "github_repo_traffic.data:main"
repo-readme = "github_repo_traffic.readme:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
